Add Fleet Manager for the web-based viewing of Dashcam footage and Error Logs
This commit introduces Fleet Manager that is a new tool for viewing Dashcam Footage and Error logs over the web. The changes encompass the implementation of the application using Flask and addition of several HTML templates for displaying the needed contents. The page layouts, login system, and also helpers to assist with file management and video streaming are also implemented. Furthermore, a necessary dependency, Flask, has been added to the pyproject.toml.

This change enhances the ability to view and manage media and logs, especially for large fleets of vehicles, providing a web interface to access footages and logs.

Note that this is a starting point and future work would include refining the user interface and adding more features such as downloading dashcam footage.