Remove commented out horizontal lines in UI code
This commit removes commented out code that previously inserted horizontal lines in the user interface layout of the sunnypilot settings.
These lines were commented out in previous versions presumably for design reasons and their removal cleans up the code and improves readability. It does not affect the functionality of the application.