Add GIT_ORIGIN parameter for improved script flexibility
In this commit, a GIT_ORIGIN parameter was added to the build_and_publish.sh and publish.sh scripts. This will allow different Git origins to be used during the build and publish process, improving flexibility and adaptability. Previously, the Git origin was hardcoded, which could potentially cause issues if different origins needed to be used.