Set docker-compose services to run under host UID:GID for permission consistency
This commit updates the `docker-compose.yml`. The user details for sunnypilot-clean and sunnypilot-build services are adjusted to run under the UID and GID of the host, defaulting to root (0) if not specified. This modification addresses permission related issues that could arise during the container runtime when service processes attempt to create or modify files owned by the host.

The .github/workflows/build_and_publish_private_dev-c3.yaml file is also updated to print user details for debugging and testing this change.