Add UID and GID to environment variables in workflow
This change adds both UID and GID to the environment variables in the GitHub workflow file 'build_and_publish_private_dev-c3.yaml'. Now, the user ID and group ID of the current user are being saved in the environment variables UID and GID, respectively. This was done to ensure the correct permission assignment when running scripts or programs that require user and group identification.